// src/lib/api/generated/models/computedFieldPreviewResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.2
 */
import type { ComputedFieldErrorResponse } from './computedFieldErrorResponse';

export interface ComputedFieldPreviewResponse {
  error?: null | ComputedFieldErrorResponse;
  value: unknown;
}